requirements.txt
## The following requirements were added by pip --freeze:
Django==1.7
Fabric==1.10.0
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
PyYAML==3.11
ansible==1.7.2
django-ipware==0.1.0
django-localflavor==1.0
ecdsa==0.11
paramiko==1.15.1
psycopg2==2.5.4
pycrypto==2.6.1
raven==5.1.1
requests==2.4.3
requirements==0.1
wsgiref==0.1.2
